feat: OA4Rust parity closure U1/U2/U4/U5/U8/U12/U13 (plan006)
U1: cluster_behavior_diffs.py — production CLI tool with self-test
U2: CI clustering steps + artifact upload (diff_candidates.tsv/md)
U4: 180+ SQL cast fixes (LIMIT/OFFSET ::bigint -> ::int) across 22 crates
U5: AUTH_EXEMPT_PATHS expanded 100+ read-only query endpoints
U8: Content-Type header for POST/PUT/PATCH in comparator.rs (already present)
U12: 985 list endpoints converted to java_success(Array, count, 0)
U13: Zero-test crates — mcp_server (19 tests), openapi (5 tests)

Verification: cargo check --lib passes (0 errors), all modified crate tests pass.

+++ oa4rust/crates/control/src/person.rs
@@ -138,7 +138,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id ASC LIMIT $2::bigint",
+                 ORDER BY unique_id ASC LIMIT $2::int",
                 &[&flag, &limit],
             )
             .await
@@ -150,7 +150,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id DESC LIMIT $2::bigint",
+                 ORDER BY unique_id DESC LIMIT $2::int",
                 &[&flag, &limit],
             )
             .await
