+++ oa4rust/crates/control/src/unit.rs
@@ -124,7 +124,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, parent_id, level FROM auth_unit \
                  WHERE deleted_at IS NULL AND (name > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name ASC LIMIT $2::bigint",
+                 ORDER BY name ASC LIMIT $2::int",
                 &[&flag, &count],
             )
             .await
@@ -134,7 +134,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, parent_id, level FROM auth_unit \
                  WHERE deleted_at IS NULL AND (name < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name DESC LIMIT $2::bigint",
+                 ORDER BY name DESC LIMIT $2::int",
                 &[&flag, &count],
             )
             .await
